backend/requirements.txt
# LangChain (aligned versions)
langchain
langchain-core
langchain-community
langchain-openai
langchain-ollama

# Clients (optional outside LangChain)
openai
ollama

# Backend
fastapi
uvicorn[standard]==0.30.6
PyYAML
python-dotenv

# Optional UI

pydantic>=2.7